requirements.txt
face_recognition
opencv-python
picamera2
python-sensors
gpiozero
spidev
